requirements.txt
Flask
opencv-python
numpy
scikit-learn
Pillow
rembg
